Move summary to GrailDiary
